k8s: Fix endpoints subsets behavior
The Kubernetes Endpoints resource is split into a list of subsets, each
consists of a list of addresses and a list of ports. So far, all subsets have
been merged together, creating the cross product over addresses and ports of
all subsets. This is incorrect. Each subset must have its cross product created
before being merged together to a list of backends.

Signed-off-by: Thomas Graf <thomas@cilium.io>